chapter_001/src/test/java/Сollection/SimpleSecondArrayTest.java
package Сollection;

import static org.junit.Assert.assertThat;
import static org.hamcrest.Matchers.is;

import org.junit.Test;

import java.util.ConcurrentModificationException;
import java.util.Iterator;
import java.util.NoSuchElementException;

public class SimpleSecondArrayTest {

    @Test
    public void whenAddThenGet() {
        SimpleSecondArray<String> array = new SimpleSecondArray<>();
        array.add("first");
        String rsl = array.get(0);
        assertThat(rsl, is("first"));
    }

    @Test
    public void whenAddThenIt() {
        SimpleSecondArray<String> array = new SimpleSecondArray<>();
        array.add("first");
        String rsl = array.iterator().next();
        assertThat(rsl, is("first"));
    }

    @Test(expected = IndexOutOfBoundsException.class)
    public void whenGetEmpty() {
        SimpleSecondArray<String> array = new SimpleSecondArray<>();
        array.get(0);
    }

    @Test(expected = IndexOutOfBoundsException.class)
    public void whenGetOutBound() {
        SimpleSecondArray<String> array = new SimpleSecondArray<>();
        array.add("first");
        array.get(1);
    }

    @Test(expected = NoSuchElementException.class)
    public void whenGetEmptyFromIt() {
        SimpleSecondArray<String> array = new SimpleSecondArray<>();
        array.iterator().next();
    }
}